Allow files to be uploaded to S3
I've borrowed the policy and signing code from a different Rails/S3
example[1]. I wanted to re-implement it so that I could really understand how
S3 uploads work, and so that I can attempt to extend the functionality with
the jQuery file upload plugin.

One difference between this and [1] is that I'm using the `form_tag` Rails
helper, which adds a 'utf8' hidden field to the form. I've had to update the
policy document to deal with this additional field.

[1] https://github.com/nelix/BrowserUpload