Use the jQuery File Upload plugin
I've had to replace `success_action_redirect` with `success_action_status` so
that the XmlHttpRequest returns a 201 instead of a 303. Returning a 303 causes
the Chrome developer tools to show the request as 'canceled' (even though the
file is uploaded correctly) which is confusing.

*NOTE* The above change means that you don't get a great experience if you
upload without Javascript now: You'll see an XML document rather than a
redirect back to the upload form.

I've added a class to the form so that I can target it with the `fileupload()`
function.